invoiceflow/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
anthropic==0.40.0
openpyxl==3.1.5
httpx==0.27.2
python-dotenv==1.0.1
aiofiles==24.1.0
pdfplumber==0.11.4
passlib[bcrypt]==1.7.4
bcrypt<4
itsdangerous==2.2.0
supabase==2.28.3
